MH-13187, Improve Track Stream Handling

+++ modules/inspection-service-ffmpeg/src/test/java/org/opencastproject/inspection/ffmpeg/MediaInspectionServiceImplTest.java
@@ -154,9 +154,9 @@ public class MediaInspectionServiceImplTest {
       Track newTrack = (Track) mi.enrich(track, false, Options.NO_OPTION);
 
       VideoStream[] videoStreams = TrackSupport.byType(newTrack.getStreams(), VideoStream.class);
-      assertTrue(videoStreams[0].getFrameCount().longValue() > 0);
+      assertTrue(videoStreams[0].getFrameCount() > 0);
       AudioStream[] audioStreams = TrackSupport.byType(newTrack.getStreams(), AudioStream.class);
-      assertTrue(audioStreams[0].getFrameCount().longValue() > 0);
+      assertTrue(audioStreams[0].getFrameCount() > 0);
       assertEquals(newTrack.getChecksum(), cs);
       assertEquals(newTrack.getMimeType(), mt);
       assertNotNull(newTrack.getDuration());
@@ -179,9 +179,9 @@ public class MediaInspectionServiceImplTest {
       Track newTrack = (Track) mi.enrich(track, false, Options.NO_OPTION);
 
       VideoStream[] videoStreams = TrackSupport.byType(newTrack.getStreams(), VideoStream.class);
-      assertTrue(videoStreams[0].getFrameCount().longValue() > 0);
+      assertTrue(videoStreams[0].getFrameCount() > 0);
       AudioStream[] audioStreams = TrackSupport.byType(newTrack.getStreams(), AudioStream.class);
-      assertTrue(audioStreams[0].getFrameCount().longValue() > 0);
+      assertTrue(audioStreams[0].getFrameCount() > 0);
       assertEquals(newTrack.getChecksum(), cs);
       assertEquals(newTrack.getMimeType(), mt);
       assertNotNull(newTrack.getDuration());
